+++ app/src/main/java/woowacourse/shopping/data/db/cart/CartContract.kt
@@ -1,11 +1,12 @@
-package woowacourse.shopping.data.database.product
+package woowacourse.shopping.data.db.cart
 
 import android.provider.BaseColumns
 
-object ProductContract : BaseColumns {
-    const val TABLE_NAME = "product"
+object CartContract : BaseColumns {
+    const val TABLE_NAME = "cart"
     const val TABLE_COLUMN_ID = "id"
     const val TABLE_COLUMN_IMAGE_URL = "image_url"
     const val TABLE_COLUMN_NAME = "name"
     const val TABLE_COLUMN_PRICE = "price"
+    const val TABLE_COLUMN_COUNT = "count"
 }
