+++ app/src/main/java/woowacourse/shopping/data/db/product/ProductContract.kt
@@ -1,12 +1,11 @@
-package woowacourse.shopping.data.database.cart
+package woowacourse.shopping.data.db.product
 
 import android.provider.BaseColumns
 
-object CartContract : BaseColumns {
-    const val TABLE_NAME = "cart"
+object ProductContract : BaseColumns {
+    const val TABLE_NAME = "product"
     const val TABLE_COLUMN_ID = "id"
     const val TABLE_COLUMN_IMAGE_URL = "image_url"
     const val TABLE_COLUMN_NAME = "name"
     const val TABLE_COLUMN_PRICE = "price"
-    const val TABLE_COLUMN_COUNT = "count"
 }
