+++ app/src/main/java/woowacourse/shopping/feature/product/detail/ProductDetailActivity.kt
@@ -7,8 +7,8 @@ import android.os.Bundle
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.shopping.R
-import woowacourse.shopping.data.repository.cart.source.local.CartLocalDataSourceImpl
-import woowacourse.shopping.data.repository.recentproduct.source.local.RecentProductLocalDataSourceImpl
+import woowacourse.shopping.data.datasource.cartdatasource.CartLocalDataSourceImpl
+import woowacourse.shopping.data.datasource.recentproductdatasource.RecentProductLocalDataSourceImpl
 import woowacourse.shopping.databinding.ActivityProductDetailBinding
 import woowacourse.shopping.feature.extension.showToast
 import woowacourse.shopping.feature.list.item.ProductView.CartProductItem
