+++ app/src/main/java/woowacourse/shopping/feature/product/detail/ProductDetailPresenter.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.feature.product.detail
 
-import woowacourse.shopping.data.repository.cart.source.local.CartLocalDataSourceImpl
-import woowacourse.shopping.data.repository.recentproduct.source.local.RecentProductLocalDataSourceImpl
+import woowacourse.shopping.data.datasource.cartdatasource.CartLocalDataSourceImpl
+import woowacourse.shopping.data.datasource.recentproductdatasource.RecentProductLocalDataSourceImpl
 import woowacourse.shopping.feature.list.item.ProductView.CartProductItem
 import woowacourse.shopping.feature.model.mapper.toProductDomain
